Simplify code & move topology into a separate function
