+++ back/src/network_schema.py
